tarifas.py
A prototype of a tariff calculator for a private chilean enterprise. They need fix some problems on the process of generate the proform or a "detail" of a truck trip, because they write them one by one on a word, and in times of hardwork, they collapse. So that's my idea. I would be really greatfully for something that can help me, maybe a idea, tip, fix, idk.